+++ src/test/java/net/sf/jabref/logic/exporter/GroupSerializerTest.java
@@ -15,7 +15,6 @@ import net.sf.jabref.model.groups.SearchGroup;
 import net.sf.jabref.model.groups.SimpleKeywordGroup;
 
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -45,7 +44,7 @@ public class GroupSerializerTest {
 
     @Test
     public void serializeSingleSimpleKeywordGroup() {
-        SimpleKeywordGroup group = new SimpleKeywordGroup("name", GroupHierarchyType.INDEPENDENT, "keywords", "test", false, ',');
+        SimpleKeywordGroup group = new SimpleKeywordGroup("name", GroupHierarchyType.INDEPENDENT, "keywords", "test", false, ',', false);
         List<String> serialization = groupSerializer.serializeTree(GroupTreeNode.fromGroup(group));
         assertEquals(Collections.singletonList("0 KeywordGroup:name;0;keywords;test;0;0;"), serialization);
     }